Move MCP client to a separate MCP package
